+++ Config/CMakeIncludes/master_targets.cmake
@@ -4,38 +4,17 @@
 
 # add master target
 add_library(Tasmanian_master INTERFACE)
+target_link_libraries(Tasmanian_master INTERFACE Tasmanian_addons)
+install(TARGETS Tasmanian_master EXPORT "${Tasmanian_export_name}")
 
 # add :: interface, useful when using with add_subdirectory()
 add_library(Tasmanian::Tasmanian INTERFACE IMPORTED GLOBAL)
-set_target_properties(Tasmanian::Tasmanian PROPERTIES INTERFACE_LINK_LIBRARIES Tasmanian_master)
-
-# add Tasmanian_shared and Tasmanian_static, associate with Tasmanian::shared and Tasmanian::static
-foreach(_tsglibtype ${Tasmanian_libs_type})
-    if (TARGET Tasmanian_libdream_${_tsglibtype})
-        add_library(Tasmanian_${_tsglibtype} INTERFACE)
-        target_link_libraries(Tasmanian_${_tsglibtype} INTERFACE "Tasmanian_libdream_${_tsglibtype};Tasmanian_addons")
-        install(TARGETS Tasmanian_${_tsglibtype} EXPORT "${Tasmanian_export_name}")
-
-        add_library(Tasmanian::${_tsglibtype} INTERFACE IMPORTED GLOBAL)
-        set_target_properties(Tasmanian::${_tsglibtype} PROPERTIES INTERFACE_LINK_LIBRARIES Tasmanian_${_tsglibtype})
-    endif()
-    if (TARGET Tasmanian_libfortran90_${_tsglibtype})
-        add_library(Tasmanian::Fortran::${_tsglibtype} INTERFACE IMPORTED GLOBAL)
-        target_link_libraries(Tasmanian::Fortran::${_tsglibtype} INTERFACE Tasmanian_libfortran90_${_tsglibtype})
-        target_link_libraries(Tasmanian::Fortran::${_tsglibtype} INTERFACE Tasmanian_libfortran03_${_tsglibtype})
-    endif()
-endforeach()
-unset(_tsglibtype)
-
-target_link_libraries(Tasmanian_master INTERFACE Tasmanian_${Tasmanian_lib_default})
-
-install(TARGETS Tasmanian_master EXPORT "${Tasmanian_export_name}")
+target_link_libraries(Tasmanian::Tasmanian INTERFACE Tasmanian_master)
 
 if (Tasmanian_ENABLE_FORTRAN)
     add_library(Tasmanian::Fortran INTERFACE IMPORTED GLOBAL)
-    set_target_properties(Tasmanian::Fortran PROPERTIES INTERFACE_LINK_LIBRARIES Tasmanian_libfortran90_${Tasmanian_lib_default})
-
-    set_target_properties(Tasmanian::Fortran PROPERTIES INTERFACE_LINK_LIBRARIES Tasmanian_libfortran03_${Tasmanian_lib_default})
+    target_link_libraries(Tasmanian::Fortran INTERFACE Tasmanian_libfortran90)
+    target_link_libraries(Tasmanian::Fortran INTERFACE Tasmanian_libfortran03)
 endif()
 
 # add executable that has the sole purpose of testing the master target